The influence on the performance of the C optimization flag (-O) can be
significant as shown by the output of "time xtest":
```
      | QEMUv8         | HiKey960 (A73 cores only)
 -----+----------------+--------------------------
  -O0 | 2m 54s, 2m 49s | 42.28s, 42.07s
  -Os | 2m 03s, 2m 03s | 25.57s, 25.60s
  -O2 | 1m 36s, 1m 35s | 24.01s, 23.93s
```
This commit introduces CFG_CC_OPT_LEVEL (default: s) which may be set to
0, s, 2 or any value accepted by the compiler. This gives better
flexibility to chose the best level depending on the use case.

Signed-off-by: Jerome Forissier <jerome@forissier.org>

<!--
    If you are new to submitting pull requests to OP-TEE, then please have a
    look at the list below and tick them off before submitting the pull request.

    1. Read our contribution guidelines:
         https://optee.readthedocs.io/en/latest/general/contribute.html

    2. Read the contribution section in Notice.md and pay extra attention to the
       "Developer Certificate of Origin" in the contribution guidelines.

    3. You should run checkpatch preferably before submitting the pull request.

    4. When everything has been reviewed, you will need to squash, rebase and
       add tags like `Reviewed-by`, `Acked-by`, `Tested-by` etc. More details
       about this can also be found on the link provided above.

    NOTE: This comment will not be shown in the pull request, so no harm keeping
    it, but feel free to remove it if you like.
-->