@jforissier 2020-09-09
I noticed that `CFG_CC_OPT_LEVEL=3` doesn't work. If anyone feels like debugging optimized code... :wink: